requirements.txt
flask>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
python-dotenv>=0.19.0
loguru>=0.6.0
joblib>=1.0.0
gunicorn>=20.0.0
pyyaml
flask-cors
jupyter==1.0.0
notebook==6.5.4
seaborn==0.12.2
matplotlib==3.7.1
ipywidgets==8.0.6